Adding docstring and refactoring

+++ src/clab_vm_startup/vms/vr.py
@@ -36,10 +36,18 @@ LOGGER = logging.getLogger(__name__)
 
 
 class VirtualRouter:
+    """
+    This class represents a virtual router.
+    """
 
     NICS_PER_PCI_BUS = 26  # Tested to work with Xrv
     NIC_TYPE = "e1000"
     TFTP_FOLDER = Path("/tftpboot")
+
+    QEMU_MONITOR_PORT = 4000  # Port to open for the qemu monitor
+
+    SERIAL_CONSOLE_PORT = 5000  # Port offset for the serial consoles
+    SERIAL_CONSOLE_COUNT = 1  # Number of serial console to open
 
     def __init__(
         self,
@@ -52,6 +60,16 @@ class VirtualRouter:
         mgmt_nic_type: Optional[str] = None,
         forwarded_ports: Optional[List[PortForwarding]] = None,
     ) -> None:
+        """
+        :param host: The host on which we deploy this VM
+        :param connection: The type of connection to setup between the host and the vm
+        :param disk_image: The path to the disk image of the vm
+        :param vcpus: The number of virtual cpus to give to the vm
+        :param ram: The amount of ram (MB) to give to the vm
+        :param nics: The amount of network interface to attach to the vm
+        :param mgmt_nic_type: The type of the mgmt interface we create and attach to the vm
+        :param forwarded_ports: A list of forwarded ports going from the host to the vm
+        """
         self.host = host
         self.connection = connection
         self.disk_image = disk_image
@@ -66,14 +84,19 @@ class VirtualRouter:
         self._boot_process: Optional[subprocess.Popen] = None
         self._stdout_thread: Optional[threading.Thread] = None
         self._stderr_thread: Optional[threading.Thread] = None
-        self._qemu_monitor: Optional[Telnet] = None
 
     @property
-    def running(self) -> bool:
-        return self._boot_process is not None and self._boot_process.returncode is None
+    def started(self) -> bool:
+        """
+        Whether this vm has already been started and not stopped
+        """
+        return self._boot_process is not None
 
     @property
     def _mgmt_interface_boot_args(self) -> List[Tuple[str, str]]:
+        """
+        Qemu boot arguments to add the management interface to the vm
+        """
         hostfwd = ""
         for forwarded_port in self.forwarded_ports:
             hostfwd += (
@@ -88,6 +111,9 @@ class VirtualRouter:
 
     @property
     def _nics_boot_args(self) -> Sequence[Tuple[str, str]]:
+        """
+        Qemu boot arguments to add the interfaces to the vm
+        """
         qemu_args = []
 
         for nic_index in range(1, self.nics + 1):
@@ -117,6 +143,9 @@ class VirtualRouter:
 
     @property
     def boot_args(self) -> List[str]:
+        """
+        Qemu boot arguments
+        """
         qemu_args = [
             ("qemu-system-x86_64",),
             ("-display", "none"),
@@ -124,7 +153,7 @@ class VirtualRouter:
             ("-m", str(self.ram)),
             ("-cpu", "host"),
             ("-smp", f"cores={self.vcpus},threads=1,sockets=1"),
-            ("-monitor", "tcp:0.0.0.0:4000,server,nowait"),
+            ("-monitor", f"tcp:0.0.0.0:{self.QEMU_MONITOR_PORT},server,nowait"),
             ("-drive", f"if=ide,file={self.disk_image}"),
         ]
 
@@ -143,6 +172,15 @@ class VirtualRouter:
         # Setup nics args
         qemu_args.extend(self._nics_boot_args)
 
+        # Setup serial consoles
+        for i in range(0, self.SERIAL_CONSOLE_COUNT):
+            qemu_args.append(
+                (
+                    "-serial",
+                    f"telnet:0.0.0.0:{self.SERIAL_CONSOLE_PORT + i},server,nowait",
+                )
+            )
+
         return [elem for arg in qemu_args for elem in arg]
 
     @abstractmethod
@@ -152,6 +190,9 @@ class VirtualRouter:
         """
 
     def _start(self) -> None:
+        """
+        Start the VM and the port forwarding processes
+        """
         # Starting vm process
         boot_args = self.boot_args
         boot_command = " ".join(boot_args)
@@ -165,15 +206,12 @@ class VirtualRouter:
             executable="/bin/bash",
         )
 
-        def stop_logging() -> bool:
-            return not self.running
-
+        # Starting logging threads
         self._stdout_thread = threading.Thread(
             target=io_logger,
             args=(
                 self._boot_process.stdout,
                 f"{boot_args[0]}[{self._boot_process.pid}]-stdout",
-                stop_logging,
             ),
         )
         self._stdout_thread.start()
@@ -183,29 +221,13 @@ class VirtualRouter:
             args=(
                 self._boot_process.stderr,
                 f"{boot_args[0]}[{self._boot_process.pid}]-stderr",
-                stop_logging,
             ),
         )
         self._stderr_thread.start()
 
-        # Connecting to qemu monitor
-        max_retry = 5
-        for _ in range(0, max_retry):
-            try:
-                self._qemu_monitor = Telnet("127.0.0.1", 4000)
-                LOGGER.debug("Successfully connected to QEMU monitor")
-                break
-            except ConnectionRefusedError:
-                pass
-
-            time.sleep(1)
-
-        if self._qemu_monitor is None:
-            raise RuntimeError(f"Failed to connect to QEMU monitor after {max_retry} attempts")
-
         LOGGER.debug("Starting port forwarding processes")
         for forwarded_port in self.forwarded_ports:
-            if forwarded_port.enabled:
+            if forwarded_port.started:
                 LOGGER.warning(f"The following port forwarding process should not be enabled but is: `{forwarded_port.cmd}`")
                 continue
 
@@ -221,7 +243,7 @@ class VirtualRouter:
         """
         Start the VM, if it hasn't been started before.
         """
-        if self.running:
+        if self.started:
             raise RuntimeError("Can not start the vm, it is already running")
 
         LOGGER.debug("Ready to start VM")
@@ -239,7 +261,7 @@ class VirtualRouter:
         LOGGER.debug("Calling pre-start")
         self.pre_start()
 
-        # Starting the vm and the qemu monitor console
+        # Starting the vm
         self._start()
 
         # Running post-start step, any vm class inheriting from this class can run extra step in there
@@ -258,22 +280,24 @@ class VirtualRouter:
 
     def _stop(self) -> None:
         """
-        Stops the VM and close the telnet connection
+        Stops the VM and the port forwarding processes
         """
         LOGGER.debug("Stopping port forwarding processes")
         for forwarded_port in self.forwarded_ports:
-            if not forwarded_port.enabled:
+            if not forwarded_port.started:
                 LOGGER.warning(f"The following port forwarding process should be running but isn't: `{forwarded_port.cmd}`")
                 continue
 
             forwarded_port.stop()
 
-        if self._qemu_monitor is not None:
-            # TODO graceful shutdown with `system_powerdown`
-            self._qemu_monitor.close()
-
         if self._boot_process:
-            self._boot_process.kill()
+            if not self._boot_process.returncode:
+                self._boot_process.kill()
+
+            # Closing streams manually to let logging thread finish
+            self._boot_process.stdout.close()
+            self._boot_process.stderr.close()
+
             self._boot_process.wait(5)
             self._boot_process = None
 
@@ -297,7 +321,7 @@ class VirtualRouter:
         """
         Stop the VM, if it is running.
         """
-        if not self.running:
+        if not self.started:
             raise RuntimeError("Can not stop the vm, it is not running")
 
         LOGGER.debug("Ready to stop VM")
@@ -317,3 +341,53 @@ class VirtualRouter:
         stop_time = time.time()
         shutdown_duration = datetime.timedelta(seconds=stop_time - start_time)
         LOGGER.info(f"VM was successfully shutdown in {str(shutdown_duration)}")
+
+    def get_qemu_monitor_connection(self) -> Telnet:
+        """
+        Get a telnet object with an open connection to the qemu monitor.
+        The caller of this method has the responsability of closing the connection.
+        """
+        LOGGER.debug("Opening connection to qemu monitor")
+        # Connecting to qemu monitor
+        max_retry = 5
+        for _ in range(0, max_retry):
+            try:
+                qemu_monitor = Telnet("127.0.0.1", self.QEMU_MONITOR_PORT)
+                LOGGER.debug("Successfully connected to QEMU monitor")
+                return qemu_monitor
+            except ConnectionRefusedError:
+                pass
+
+            time.sleep(1)
+
+        raise RuntimeError(f"Failed to connect to QEMU monitor after {max_retry} attempts")
+
+    def get_serial_console_connection(self, index: int = 0) -> Telnet:
+        """
+        Get a telnet object with an open connection to a serial console of the vm.
+        The caller of this method has the responsability of closing the connection.
+        The caller can specify an index if multiple serial console have been created.
+            The index starts at zero.
+
+        :param index: The index of the serial console to get.
+        """
+        if index >= self.SERIAL_CONSOLE_COUNT:
+            raise ValueError(
+                "The provided serial console index exceeds the highest one we created: "
+                f"{index} > {self.SERIAL_CONSOLE_COUNT - 1}"
+            )
+
+        LOGGER.debug(f"Opening connection to serial console (with index={index})")
+        # Connecting to qemu monitor
+        max_retry = 5
+        for _ in range(0, max_retry):
+            try:
+                xr_console = Telnet("127.0.0.1", self.SERIAL_CONSOLE_PORT + index)
+                LOGGER.debug("Successfully connected to serial console")
+                return xr_console
+            except ConnectionRefusedError:
+                pass
+
+            time.sleep(1)
+
+        raise RuntimeError(f"Failed to connect to serial console after {max_retry} attempts")
